backend/services/document_converter/pyproject.toml
[project]
name = "document-converter"
version = "0.1.0"
description = "ePUB to structured JSON conversion service for Readwell"
requires-python = ">=3.10"
dependencies = [
    "ebooklib>=0.17",
    "beautifulsoup4>=4.9",
    "Pillow>=10.0",
    "lxml>=4.6",
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "boto3>=1.26",
    "click>=8.0",
    "tqdm>=4.50",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
document-converter = "document_converter.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["document_converter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
